+++ suite-common/wallet-core/src/settings/walletSettingsConstants.ts
@@ -6,6 +6,7 @@ const SET_BITCOIN_AMOUNT_UNITS = '@suite/set-bitcoin-amount-units';
 const TOGGLE_HIDE_SUSPICIOUS_TRANSACTIONS = '@wallet-settings/toggle-hide-suspicious-transactions';
 const SET_DISCREET_MODE = '@wallet-settings/set-discreet-mode';
 const CHANGE_COIN_VISIBILITY = '@wallet-settings/change-coin-visibility';
+const SET_MEV_PROTECTION = '@wallet-settings/set-mev-protection';
 
 export const WALLET_SETTINGS = {
     SET_BASE_CURRENCY,
@@ -16,4 +17,5 @@ export const WALLET_SETTINGS = {
     TOGGLE_HIDE_SUSPICIOUS_TRANSACTIONS,
     SET_DISCREET_MODE,
     CHANGE_COIN_VISIBILITY,
+    SET_MEV_PROTECTION,
 } as const;