+++ packages/suite/src/storage/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 25.9.0
 
 - persist `suite.stakingDashboardCollapsed` in `suiteSettings`
+- persist `wallet.settings.mevProtection` in `walletSettings`
 
 ## 25.8.0